Move some features to 'game_fixes' folder
