+++ spt/features/game_fixes/bms_flashlight_fix.cpp
