+++ spt/features/game_fixes/hardlock.cpp
