+++ spt/features/game_fixes/multi_instance.cpp
