+++ spt/features/game_fixes/noclip_nofix.cpp
