+++ spt/features/game_fixes/rng.cpp
