+++ spt/features/game_fixes/rng.hpp
