+++ spt/features/game_fixes/vag_crash.cpp
