+++ spt/features/game_fixes/visual_fixes.cpp
